Allow place_param2 = 0 node placement predictions
The placement prediction value 0 was accidentally ignored
and made the clients fall back to automatic rotation based
on the node paramtype2 value.

This now changes the internal representation to s16 to properly
indicate the disabled state (e.g. 'nil' in Lua).

+++ src/itemdef.h
@@ -87,7 +87,7 @@ struct ItemDefinition
 	// Server will update the precise end result a moment later.
 	// "" = no prediction
 	std::string node_placement_prediction;
-	u8 place_param2;
+	s16 place_param2; // -1: not defined, >= 0: hardcoded prediction
 
 	/*
 		Some helpful methods
